+++ app/src/main/kotlin/io/treehouses/remote/Fragments/HomeFragment.kt
@@ -50,7 +50,7 @@ class HomeFragment : BaseHomeFragment(), SetDisconnect {
     private lateinit var bind: ActivityHomeFragmentBinding
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         bind = ActivityHomeFragmentBinding.inflate(inflater, container, false)
-        mChatService = listener.getChatService()
+//        mChatService = listener.getChatService()
         mBluetoothAdapter = BluetoothAdapter.getDefaultAdapter()
         preferences = PreferenceManager.getDefaultSharedPreferences(context)
         setupProfiles()
@@ -164,7 +164,7 @@ class HomeFragment : BaseHomeFragment(), SetDisconnect {
     }
 
     override fun checkConnectionState() {
-        mChatService = listener.getChatService()
+//        mChatService = listener.getChatService()
         if (mChatService.state == Constants.STATE_CONNECTED) {
             showLogDialog(preferences!!)
             transition(true, arrayOf(150f, 110f, 70f))
